requirements.txt
python-fasthtml
mysql-connector-python
fastapi